Fixes some nullref errors
comments out section in EnemyWalkBehavour
changes tutorial enemy
fixes missing script on player and gamemanager

+++ Assets/Scripts/Combat_Refactored/Actor.cs
@@ -6,6 +6,6 @@ using UnityEngine;
 [RequireComponent(typeof(Rigidbody2D))]
 public class Actor : MonoBehaviour
 {
-    public virtual void TakeDamage(int amount)
+    public virtual void TakeDamage(float amount)
     { }
 }
