+++ Assets/Scripts/Combat_Refactored/Enemy/Enemy.cs
@@ -110,7 +110,7 @@ public class Enemy : Actor, IHaveHealth
     /// This function is called when the hurtbox collides with a hitbox, it reduces the health of the object the hurtbox belongs to by the value of the int amount
     /// </summary>
     /// <param name="amount">value to reduce health by when function is called, the amount belongs to the hitbox that collides with the hurtbox</param>
-    public override void TakeDamage(int amount)
+    public override void TakeDamage(float amount)
     {
         //this runs to make sure Health doesn't fall into the negatives
         if (Health <= 0)
